fix: close r4 release rebase lifecycle races
Agend-Agent: archfix-codex-dev
Agend-Task: t-20260714044556193442-27634-5
Agend-Branch: fix/release-transaction-s2-force-rebase
Agend-Issued-At: 2026-07-14T08:21:06.719016+00:00
